move type coercer to scheam.py
