+++ python/sparktk/frame/frame.py
@@ -19,7 +19,7 @@ from pyspark.rdd import RDD
 from pyspark.sql import DataFrame
 
 from sparktk.frame.pyframe import PythonFrame
-from sparktk.frame.schema import schema_to_python, schema_to_scala
+from sparktk.frame.schema import schema_to_python, schema_to_scala, schema_is_coercible
 from sparktk import dtypes
 import logging
 logger = logging.getLogger('sparktk')
@@ -106,7 +106,7 @@ class Frame(object):
                 logger.debug("%s values were unable to be parsed to the schema's data type." % validate_schema_result.bad_value_count)
 
             # If schema contains matrix datatype, then apply type_coercer to convert list[list] to numpy ndarray
-            map_source = MatrixCoercion.schema_is_coercible(source, list(schema))
+            map_source = schema_is_coercible(source, list(schema))
             self._frame = PythonFrame(map_source, schema)
 
     def _merge_types(self, type_list_a, type_list_b):
@@ -266,7 +266,7 @@ class Frame(object):
         if self._is_python:
             # If schema contains matrix dataype,
             # then apply type_coercer_pymlib to convert ndarray to pymlib DenseMatrix for serialization purpose at java
-            self._frame.rdd = MatrixCoercion.schema_is_coercible(self._frame.rdd, list(self._frame.schema), True)
+            self._frame.rdd = schema_is_coercible(self._frame.rdd, list(self._frame.schema), True)
             # convert PythonFrame to a Scala Frame"""
             scala_schema = schema_to_scala(self._tc.sc, self._frame.schema)
             scala_rdd = self._tc.sc._jvm.org.trustedanalytics.sparktk.frame.internal.rdd.PythonJavaRdd.pythonToScala(self._frame.rdd._jrdd, scala_schema)
@@ -283,7 +283,7 @@ class Frame(object):
             python_schema = schema_to_python(self._tc.sc, scala_schema)
             python_rdd = RDD(java_rdd, self._tc.sc)
             # If schema contains matrix datatype, then apply type_coercer to convert list[list] to numpy ndarray
-            map_python_rdd = MatrixCoercion.schema_is_coercible(python_rdd, list(python_schema))
+            map_python_rdd = schema_is_coercible(python_rdd, list(python_schema))
             self._frame = PythonFrame(map_python_rdd, python_schema)
         return self._frame
 
@@ -422,68 +422,3 @@ class SchemaValidationReturn(PropertiesObject):
         """
         return self._bad_value_count
 
-
-class MatrixCoercion(object):
-    @staticmethod
-    def schema_is_coercible(source, python_schema, in_scala=False):
-        """
-        check whether python schema is coercible or not.
-        Like if schema contains matrix datatype, convert list[list] to numpy ndarray
-        """
-        flag = False
-        for schema in python_schema:
-            if type(schema[1]) == dtypes._Matrix:
-                flag = True
-                break
-
-        if flag:
-            if in_scala:
-                map_source = source.map(MatrixCoercion.type_coercer_pymllib(python_schema))
-            else:
-                map_source = source.map(MatrixCoercion.type_coercer(python_schema))
-        else:
-            map_source = source
-
-        return map_source
-
-    @staticmethod
-    def type_coercer(schema):
-        """
-        When creating a new frame(python frame created) or converting frame from scala to python frame,
-        the function scans a row and performs below
-            * when creating a new frame(python frame created) - if it finds list[list](which represents matrix) as column value,
-              converts it to numpy ndarray
-            * when Converting frame from scala to python frame - (scala converts DenseMatrix--> JList[JList[Double]](in JConvert.scala),
-              jconvert.py converts JList[JList[Double]] --> list[list[float64]]), converts list[list] to ndarray
-
-        """
-        def decorator(row):
-            import numpy as np
-            for i in xrange(len(schema)):
-                if type(schema[i][1]) == dtypes._Matrix:
-                    if isinstance(row[i], list):
-                        row[i] = np.array(row[i], dtype=np.float64)
-            return row
-        return decorator
-
-    @staticmethod
-    def type_coercer_pymllib(schema):
-        """
-        When converting from python to scala, function scans the row and converts the ndarray
-        to python mllib DenseMatrix. so that autopicklers understands how to serialize from pyspark mllib DenseMatrix to Scala MLlib DenseMatrix.
-        For Serialization to work we have to explicitly call SparkAliases.getSparkMLLibSerDe in pythonToScala() method of PythonJavaRdd.scala class
-
-        ndarray stores data as row-major where as mllib densematrix stores data as column-major.
-        To construct mllib DenseMatrix with row-major we are setting isTransposed=True.
-        """
-        def decorator(row):
-            from pyspark.mllib.linalg import DenseMatrix
-            for i in xrange(len(schema)):
-                if type(schema[i][1]) == dtypes._Matrix:
-                    shape = row[i].shape
-                    arr=row[i].flatten()
-                    # By default Mllib DenseMatrix constructs column-major matrix.
-                    # Setting isTranposed=True, will construct row-major DenseMatrix
-                    row[i] = DenseMatrix(shape[0], shape[1], arr, isTransposed=True)
-            return row
-        return decorator
